% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Taylor Swift Dashboard App}
\usage{
launch_app()
}
\description{
This function launches the Shiny dashboard for exploring Taylor Swift's albums and songs.
}
