% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taylor_album}
\alias{taylor_album}
\title{Taylor Swift Albums}
\format{
A data frame with Taylor Swift albums.
}
\usage{
data(taylor_album)
}
\description{
Taylor Swift Albums
}
\keyword{datasets}
