% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taylor_album_songs}
\alias{taylor_album_songs}
\title{Taylor Swift Album Songs}
\format{
A data frame with song details from albums.
}
\usage{
data(taylor_album_songs)
}
\description{
Taylor Swift Album Songs
}
\keyword{datasets}
