% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taylor_all_songs}
\alias{taylor_all_songs}
\title{All Taylor Swift Songs}
\format{
A data frame with Taylor Swift's songs and related information.
}
\usage{
data(taylor_all_songs)
}
\description{
All Taylor Swift Songs
}
\keyword{datasets}
