% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylorswift13-package.R
\docType{package}
\name{taylorswift13-package}
\alias{taylorswift13}
\alias{taylorswift13-package}
\title{taylorswift13: Taylor Swift Lyrics and Song Data for Taylor Swift's Discography}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides comprehensive datasets and analysis tools for Taylor Swift's albums and songs, including an interactive Shiny app for data visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://etc5523-2024.github.io/taylorswift13/}
}

}
\author{
\strong{Maintainer}: Shalini Nair \email{snai0031@student.monash.edu}

}
\keyword{internal}
