% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taylor_album}
\alias{taylor_album}
\title{Taylor Swift Albums}
\format{
A data frame with the following columns:
\describe{
\item{X}{Row identifier}
\item{album_name}{Name of the album}
\item{ep}{Indicates whether the album is an EP (Extended Play)}
\item{album_release}{Release date of the album}
\item{metacritic_score}{Metacritic aggregate score for the album}
\item{user_score}{User-generated score for the album}
}
}
\usage{
taylor_album
}
\description{
A data frame listing Taylor Swift's albums.
}
\examples{
data(taylor_album)
head(taylor_album)
summary(taylor_album)
}
\keyword{datasets}
